requirements.txt
Werkzeug
mysql-connector-python
Flask-Limiter
requests
redis
flask
gunicorn
python-dotenv
celery


